backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0

# PDF Generation
reportlab==4.1.0
Pillow==10.2.0

# Validation
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# Security
python-dotenv==1.0.1

# Redis (optional for caching)
redis==5.0.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0

# Utilities
python-dateutil==2.8.2
